core: limit disk alias and description for disk metadata
On block storage, the disk metadata is limited to 512 bytes.
Since we store its alias and description in the metadata, we should
limit their size or else they will be truncated by vdsm and we won't be
able to read them.
This patch limits the value of the metadata description field size to a
maximum of 210 bytes. We do it also for file storage in order to prevent
a truncation of the description field when moving a disk from file to
block storage.

Change-Id: I74d04291bbaacb5e642d8658ab752b1a8d32c0ac
Bug-Url: https://bugzilla.redhat.com/1258547
Related-To: https://bugzilla.redhat.com/1249130
Related-To: https://bugzilla.redhat.com/1261387
Signed-off-by: Idan Shaby <ishaby@redhat.com>