restapi: Remove the capability to pin a VM to a single host
In version 3.6 the capability to pin a VM to multiple hosts was added to
the RESTAPI, using the "hosts" element:

  PUT /vms/{vm:id}
  <vm>
    <placement_policy>
      <hosts>
        <host id="{host:id}/">
        <host id="{host:id}/">
        ...
      </hosts>
    </placement_policy>
  </vm>

The existing capability to pin to a single host was preserved, for
backwards compatibility. This patch removes that capability. It is
still possible to pin to a single host specifying a list of hosts with
one element:

  PUT /vms/{vm:id}
  <vm>
    <placement_policy>
      <hosts>
        <host id="{host:id}/">
      </hosts>
    </placement_policy>
  </vm>

Change-Id: I4e5de031b232e60523ff05088c848a5dab5c185c
Bug-Url: https://bugzilla.redhat.com/1260732
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>