core: Update the version of Jandex
This patch updates the version of Jandex used by the JBoss modules
plugin, so that it is the same actually used by the application server.
In addition this updated version of Jandex fixes a bug that forced us to
do an extra copy of .jar files. We no longer need to do that.

Change-Id: I1eb62c51e7199483c1869f3eff3244fba2b1b6a6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>