core: spmStart failure when clusterlock inquire isn't supported
As part of the SPM selection flow in the engine we run getSpmStatus()
on the candidate host to get the current spm status (if possible) and
we act according to the result.

After an "offline" (when the dc is down) DC upgrade from DC that uses
V1 for it's domain (3.0 for example) to a version that uses the
StoragePoolMemoryBackend we might fail when attempting to get the current
spm status, that's because StoragePoolMemoryBackend relies on the fact
that the used clusterlock supports inquiring while it might not.

This patch ignores the failure of getSpmStatus() if the failure is
InquireNotSupportedError and attempts to start the spm, if the cluster
lock is acquired we'll fail on the spmStart flow.

Change-Id: I996fc4454c210dd278dafa9d94a13c4981162f5b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1242092
Signed-off-by: Liron Aravot <laravot@redhat.com>