core: fix graphics on import
1. restore display type values:
the values for cirrus (vnc) and qxl changed, this breaks backward
compatibility with previously exported vms, so restored to original
values.

2. the code that create graphics device on import was called too soon,
and device was removed while reloading the vm from export domain during
canDoAction, so moved the call to a better place.

3. fixed that code to select spice for qxl, if os support it, to match
the previous values in the engine (qxl was always spice)

Change-Id: I3fd8351a3f18428cfe905a39221756608b64da05
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1261446
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>