webadmin: there are no such things as ghost in MacRangeWidget
'isGhost' mechanism is faulty. Correct behavior is, acording to
colleges and code notes, that there's at most one ghost and it has to
be the last record. Currently implementation allows existence of
multiple ghosts, which needs not to be trailing records.

For simplicity, in this dialog we assume that there cannot exist
ghosts, until ghost problem is solved generally

Change-Id: I8091cedfd3cf51dfcb93bb7c7de45234088bfdd8
Bug-Url: https://bugzilla.redhat.com/1217940
Signed-off-by: Martin Mucha <mmucha@redhat.com>