restapi: Remove the "timezone" element
Currently the VM resource uses "timezone" element to represent the time
zone. This element only allows a string:

  <vm>
    <timezone>Europe/Madrid</timezone>
  </vm>

This doesn't allow extension, and as a it was necessary to add the UTC
offset, it was replaced with a new structured `time_zone` element:

  <vm>
    <time_zone>
      <name>Europe/Madrid</name>
      <utc_offset>GMT+1</utc_offset>
    </time_zone>
  </vm>

The old `timezone` element was preserved for backwards compatibility,
but this patch removes it completely.

Change-Id: Idbd95d7f0f70fb22b604644196b88b071f027060
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>