restapi: Remove the data center "storage_type" element
Data centers used to be associated to a specific storage type (NFS,
Fiber Channel, iSCSI, etc) but they have been changed some time so that
there are only two types: with local storage and with shared storage. A
new "local" element was introduced to indicate this, and the old
"storage_type" was element was preserved for backwards compatibility.
This patch removes this old element completely.

Change-Id: If07b8bc8e32c340df5665c98b324920a8e1a006e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>