jsonrpc: set the protocol as default
We want to phase out xmlrpc so for cluster levels 3.6+ we won't let the
user to use xmlrpc and set jsonrpc as default. Whenever we upgrade
version of cluster we forbid when there are host which still use xmlrpc.

As part of this patch we remove protocol which was never implemented and
there are no plans to provide implementation.

Change-Id: I34dbd73ea5d79a002ac5714f41b910f7e171daf0
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Bug-Url: https://bugzilla.redhat.com/1249374