engine : CoCo infrastructure should provide a timeout for each command
A new vdc_option in the database CoCoCommandLifeInMinutes
determines the lifetime of a command. A command that is
still in Active status after the lifetime is marked as Failed.
The onFailed method is invoked on the command's callback which
is responsible for releasing any locks on the entities.

Commands can fine tune the life time of a command by
invoking VdcActionParametersBase.setLifeInMinutes

Change-Id: I984f06d05f1ac43985cb9a8bda6dad7cfa5ed3ef
Bug-Url: https://bugzilla.redhat.com/1261733
Signed-off-by: Ravi Nori <rnori@redhat.com>