core: verify image removal during Live Merge
In the multi-step Live Merge process, an image is removed after the
merge operation is completed by use of the deleteVolume vdsm verb.  If
vdsm dies during this task the task must be recovered when vdsm starts
up, resulting in a task status of 'cleanSuccess'.  Unfortunately, this
status is ambiguous regarding whether the deletion succeeded or not, and
this causes the entire Live Merge operation to fail.  Not only does the
initial attempt fail in this case, but a retry will encounter the same
error and cleaning up the system requires manual intervention.

This fix adds another step to Live Merge to check whether image deletion
succeeds regardless of the returned status of deleteVolume.  This allows
Live Merge to succeed in above case where deletion did occur yet the
command was unable to report the success.

Note that SPM removal will fix this problem because deleteVolume will no
longer report any ambiguous statuses, so this fix may be reverted once
SPM removal is in place and the engine no longer needs compatibility
with host versions having SPM.

Change-Id: Ib3d8e35ff65cf2d9e0cf141a8a20f5dcd6cf8544
Bug-Url: https://bugzilla.redhat.com/1227497
Signed-off-by: Greg Padgett <gpadgett@redhat.com>