core: Support detach of local Storage Domain
When detaching a local Storage Domain, the engine will automatically try
to remove it.
The remove process in the engine, checks if local Storage
Domain is already attached to a Data Center so the Host will not fail
during the remove process.

This validation is meanningless for local Storage Domain since the local
Storage Domain will first get detached and only after it will be removed
as part of the Remove process.

The detach operation should remove its storage pool metadata, and that
is why the local Storage Domain should not be obligated to this validation.

Change-Id: I9e4e603301a3ec893e6de5804a11d136dd1965ec
Bug-Url: https://bugzilla.redhat.com/1263589
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>