core: add missing audit log on migration failure
In the following scenario:
1. Migrate VM
2. VM fails to migrate to the first host
3. There are no other hosts to migrate to

Assuming the number of retries for a VM migration is not
set to 1, the audit log VM_MIGRATION_FAILED was not
generated but VM_MIGRATION_FAILED_NO_VDS_TO_RUN_ON was
generated instead.

This patch fixes it by changing the audit log
VM_MIGRATION_FAILED_NO_VDS_TO_RUN_ON to be a warning
that only specifies that no host was found and after it
VM_MIGRATION_FAILED is generated to summarize that the
migration failed.

Change-Id: I8aaa2004102269fed78eb737a0c1806821e711ef
Bug-Url: https://bugzilla.redhat.com/1193025
Signed-off-by: Arik Hadas <ahadas@redhat.com>