webadmin: Bonding modes 0, 5 and 6 should be avoided for VM networks
Bonds with bonding modes 0, 5 and 6 should be prohibited from
having vm networks.
This patch modifies the setup networks dialog not to allow to
attach a vm network to a bond with such bonding options, and also
prohibits to change the bonding options to such if a vm network
is attached.

Change-Id: I9bc5f0da68ce8cc55bda1c7aaf8568e7bbd9bda0
Bug-Url: https://bugzilla.redhat.com/1094842
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>