webadmin: Fixed remove storage qos error popup
When removing storage QoSes(Quality Of Services) which had disk profiles
attached to them, together with QoSes that didn't, the confirmation
dialog would show only the QoSes with the disk profiles.

The fix changed that and now the confirmation dialog is showing all of
the QoSes that are about to be removed.

Also, Style can be added to the list items in the "Remove Qos Model"
using html tagging.

Change-Id: I66d7bd71ed736d61c72deb885b8a37c35e95f030
Bug-Url: https://bugzilla.redhat.com/1142865
Signed-off-by: Tomer Saban <tsaban@redhat.com>