core: Add custom findbugs plugin
The version of the findbugs plugin that we currently use (version 2.5.2)
doesn't work correctly with Java 8, and new versions of the plugin don't
support multiple filters files, which we use everywhere. To avoid this
issue this patch introduces a custom Maven plugin that merges the
multiple filters file into one during runtime, and then executes the
findbugs plugin passing only one file. This way we can migrate to the
new version of find bugs and then to Java 8.

This patch also includes a "findbugs.sh" script that contains the
command required to run the custom plugin.

Change-Id: If136b2bbaeb0d53017ee8322c76772a226045f55
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>