core: Filter out external VMs for OVF_SOTRE disk.
Filter out external VMs when we store VMs to the OVF_STORE disk in the
Storage Domain.

External VMs represent processes which run on the Host and does not
reflect entities which the user will want to preserve for disaster,
therefore there is no need to preserve also external VMs in the
OVF_STORE disk.

Change-Id: I253bfb974f78d0e7441da020d4bfe46c8aa5ac10
Bug-Url: https://bugzilla.redhat.com/1259467
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>