core: Remove BackwardCompatibilityTaskCreationTest
This test was added in order to future-proof the code against
accidentally changing the behavior of concreteCreateTask() as part of
the development of the SEAT framework.

As this framework is now deprecated in favor of Command Coordination,
and this method does nothing but delegate to CommandCoordinationUtil,
there really is no reason to keep this test. Especially in light of the
occasional failures it displayed in the CI environment.

Change-Id: I56b1f6378d3b9e6df9cfaaefc1d2bad260bffaf9
Bug-Url: https://bugzilla.redhat.com/1227310
Signed-off-by: Allon Mureinik <amureini@redhat.com>