frontend: add warning when taking a snapshot with memory
Currently when taking a live snapshot of a VM the memory saving is not
"live", i.e the VM is paused for the duration of dumping the memory.
Therefore, adding a warning that is going to be presented when asked
to take a live snapshot with memory which says that the VM is going
to be paused while saving the memory.

Change-Id: Id436f462b56099c1d5e9f10d99b5647a95119a9a
Bug-Url: https://bugzilla.redhat.com/1057009
Signed-off-by: Arik Hadas <ahadas@redhat.com>