core: Correctly finish Remove VM Pool job
For every VM in the pool removed, the endAction() of the corresponding
RemoveVmCommand must be called. For this purpose, RemoveVmCommand must
be parent command of all async tasks, not RemoveVmPoolCommand.

Also, adding IDs of all async tasks to the return value of
RemoveVmPoolCommand (RemoveVmCommand also needed to be fixed to add the
IDs to the correct list if RemoveVmCommand is executed as internal
command).

Change-Id: I49e31a498cf57dc8a1b459c79fbe4566cd3d0e89
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1257132
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>