core: Fix detection of preferred hosts for numa nodes
This fixes two things:

When adding new numa nodes through the rest api or the numa support
dialog in the host context menu, the detection of the preferred hosts
failed in AddVmNumaNodesCommand#canDoAction.

When creating or updating VMs, the first time the admin switched
to "non migratable", the detection of the preferred hosts failed.
This led to the situation where the whole VM was created, except
for the numa nodes.
When updating a VM all numa nodes were deleted and could not be added
again. So the whole numa configuration disappeared.

Bug-Url: https://bugzilla.redhat.com/1248049
Bug-Url: https://bugzilla.redhat.com/1256725
Change-Id: I1f72bbe131402f2037f6083c236864b5a2d3ac62
Signed-off-by: Roman Mohr <rmohr@redhat.com>