core: Exclude the ME_ENUM_FIELD_SETTER findbugs warning
Findbugs doesn't like that the OsRepositoryImpl.init method is public,
but we need that because it is the mechanism that we use to lazyly
initialize the OS repository.

This is the findbugs warning:

  Public enum method unconditionally sets its field

  This public method declared in public enum unconditionally sets enum
  field, thus this field can be changed by malicious code or by accident
  from another package. Though mutable enum fields may be used for lazy
  initialization, it's a bad practice to expose them to the outer world.
  Consider removing this method or declaring it package-private.

  Bug kind and pattern: ME - ME_ENUM_FIELD_SETTER

This patch changes the exclude filters for the "utils" project so that
this findbugs warning will be ignored.

Change-Id: Ibe37f727663fef92bc559edc7d2d665ccdf1aed5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>