Select the first CPU profile when multiple are available
This patch fixes a regression where a VM could not be added to
a cluster where multiple CPU profiles were defined without having
a specific CPU profile assigned by the REST call.

The fix is not ideal as we should redo the whole mechanism to make
sure all permissions and quotas are obeyed. Change like that would
be too invasive at this stage of development and we will address it
for the next version or update.

Change-Id: I462579ff12c8c4f97c72c08e12637cfedd766f10
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1262293
Signed-off-by: Martin Sivak <msivak@redhat.com>