restapi: Use "logical_units" instead of multiple "logical_unit"
Currently the logical units that are part of a volume group are reported
as an unbounded number of "logical_unit" elements. For example, when
reporting the details of a storage domain:

  GET /storagedomains/{storagedomain:id}
  <storage_domain>
    ...
    <storage>
      ...
      <volume_group>
        <logical_unit>
          <!-- First LU -->
        </logical_unit>
        <logical_unit>
          <!-- Second LU -->
        </logical_unit>
        ...
      </volume_group>
    </storage>
  </storage_domain>

This is contrary to the usual practice in the API, as list of elements
are always wrapped with an element. This has been fixed now, so the list
of logical units will be wrapped with the "logical_units" element:

  GET /storagedomains/{storagedomain:id}
  <storage_domain>
    ...
    <storage>
      ...
      <volume_group>
        <logical_units>
          <logical_unit>
            <!-- First LU -->
          </logical_unit>
          <logical_unit>
            <!-- Second LU -->
          </logical_unit>
          ...
        </logical_units>
      </volume_group>
    </storage>
  </storage_domain>

Change-Id: I08a71b3ca7873cfee6dcd0d7b881ce6bae935581
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>