restapi: Rename "GuestNicConfiguration" to "NicConfiguration"
This is to increase the consistency between the names of the complex
types and the names of the elements. In this case the name of the
element is "nic_configuration", and it is better for backwards
compatibility to change the name of the type than to change the name of
the element.

Change-Id: I4d6d2759bf2c9e6a585d8e7542fb1bf80c21fbb1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>