restapi: Don't use "TransportType" and "TransportTypes"
These two complex types are used to represent s list of strings. This
isn't worth the creation of two named complex types. This patch changes
the XML schema so it doesn't use them. The types are preserved because
they are still used by the capabilities resource, but will be removed in
the future.

Change-Id: I9f88984ec93bed0e86f6c418d7b519e84b8427b6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>