restapi: Don't use "Usage" and "Usages"
These two complex types are used to represent s list of strings. This
isn't worth the creation of two named complex types. This patch changes
the XML schema so it doesn't use them. The types are preserved because
they are still used by the capabilities resource, but will be removed in
the future.

Change-Id: I45fc6e09dd50f3ae680c4d770b18f3fcd7ac80a3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>