restapi: Fix JAXB annotations of scheduling policy unit
This complex type uses a JAXB annotation to assign to the Java getter of
the "properties" element the "getPropertiesMetaData" name. This isn't
consistent with the rest of the API, where the name of the getter uses
the plural of the name of the element, in this case "getProperties".
This patch fixes that.

Change-Id: I9d8b0520c7abff47e3e9dd5fc8bbb70c190ffd39
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>