core: Don't use Infinispan subsystem "start" attribute
The "start" attribute of the "cache-manager" element of the Infinispan
sybsystem of the application server was removed in version 9 of WildFly.
Instead of relying on that to ensure that the cache is started before
the backend it is better to explicitly require it using the EJB
deployment descriptor "ejb-jar.xml".

Change-Id: I84f2fe41a5d4e72be7a696680742599b109fad8f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>