restapi: Rename "WatchDog" to "Watchdog"
This is to increase the consistency between the names of the complex
types and the names of the elements. In this case the name of the
element is "watchdog", a single word, and it is better for backwards
compatibility to change the name of the type than to change the name of
the element.

Change-Id: I0f72725680fe849fed4b6a336c32d9fe6766b3ee
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>