restapi: Rename "CdRom" to "Cdrom"
This is to increase the consistency between the names of the complex
types and the names of the elements. In this case the name of the
element is "cdrom", a single word, and it is better for backwards
compatibility to change the name of the type than to change the name of
the element.

Change-Id: Ifb642144312401688a4ceec3749223a44fdd53ea
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>