core: Force version 3.x of findbugs
This patch changes the findbugs configuration so that instead of using a
custom plugin it uses an artifact containing the shared filters. This
artifact is then added to the class path of the findbugs plugin,
together with version 3.x of findbugs. This allows us to use the old
version of the plugin (which supports multiple filter files) with the
new version of findbugs itself (which supports Java 8).

In the future, we should be able to use the new version of the plugin,
when it eventually supports multiple filter files:

  Support multiple include and exclude filter files
  https://github.com/gleclaire/findbugs-maven-plugin/pull/28

Change-Id: Idf0d7202c125320188280ffc9eae1ba36c785fc2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>