restapi: Generate JAX-RS interfaces from model
Currently the operation model of the RESTAPI is described by the JAX-RS
interfaces, which are written manually. For exmple, the bookmark interface
for the bookmark resource is defined as follows:

  @Produces({ApiMediaType.APPLICATION_XML, ApiMediaType.APPLICATION_JSON})
  public interface BookmarkResource {
    @GET
    Bookmark get();

    @PUT
    @Consumes({ApiMediaType.APPLICATION_XML, ApiMediaType.APPLICATION_JSON})
    Bookmark update(Bookmark bookmark);

    @DELETE
    Response remove();
 }

This patch moves this operational definition to the model, where they are
also written manually, but using the meta-model language. For example:

  @Service
  public interface BookmarkService {
    interface Get {
        @Out Bookmark bookmark();
    }

    interface Update {
        @In @Out Bookmark bookmark();
    }

    interface Remove {
    }
  }

Each service is defined by an interface marked with the "@Service"
annotation. Operations are defined using nested interfaces. Parameters
defined as methods with the @In and @Out, to indicate the direction
of the parameter (input, output or both).

Some services also have "locators", methods that return references to
other services. For example, currently the bookmarks resource is defined
as follows:

  @Produces({ApiMediaType.APPLICATION_XML, ApiMediaType.APPLICATION_JSON})
  public interface BookmarksResource {
    @GET
    Bookmarks list();

    @POST
    @Consumes({ApiMediaType.APPLICATION_XML, ApiMediaType.APPLICATION_JSON})
    Response add(Bookmark bookmark);

    @Path("{id}")
    BookmarkResource getBookmarkResource(@PathParam("id") String id);
  }

Using the meta-model this service will be defined as follows:

  @Service
  public interface BookmarksService {
    interface Add {
        @In @Out Bookmark bookmark();
    }

    interface List {
        @Out Bookmark[] bookmarks();
    }

    @Service BookmarkService bookmark(String id);
 }

Here the "bookmark" method represents a locator, an operation of the service
that returns a reference to another service, in this particular case the
bookmarks service has the capability to tell the user where is the service
that manages a specific bookmark.

These new service interfaces aren't used at all during runtime, instead they
are used to populate an in-memory representation of the model, that can be
easily consumed. For example, if an application needs to find the names
of all the available services:

  Model model = ...;
  for (Service service : model.getServices()) {
    System.out.println(service.getName());
  }

In addition to moving the definition of the services to the model, this patch
also introduces a code generator that generates the JAX-RS interfaces from
the in-memory model. The resulting JAX-RS interfaces are almost identical
to the previous ones, so there is no need for massive changes to the
existing services implementations.

Change-Id: I65efb47a2c03b028c398ad40871b34665e6ddba8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>