core: Disable warnings about private modules
The application server marks its modules as "private" using the
"jaxb.api" property:

  <property name="jboss.api" value="private"/>

When one of these "private" modules is used by an application the server
issues the following warning during deployment of the application:

  JBAS018567: Deployment "deployment.myapp.ear" is using a private
  module ("org.apache.commons.lang:main") which may be changed or
  removed in future versions without notice.

We are well aware of this, and we don't consider it problematic.

This patch modifies the configuration of the logging subsystem so that
warnings issued for that category will be ignred.

Change-Id: I28ba11f34f1864500cf9e6f69b6765c3049cdf3e
Bug-Url: https://bugzilla.redhat.com/1269413
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>