db: executed config cleanup as upgrade script
Up to now when a config entry was deprecated, we had to remove it in the
pre-upgrade step calling fn_db_delete_config_value from 0000_config.sql.

This may be problematic if a reference to the deleted entry exists in
one
of the upgrade scripts, in such case the entry will be removed in the
pre-upgrade step and then the upgrade script referring this key will
fail
and cause all the upgrade process to fail.

This patch moves all deprecated config entries to an upgrade script
resolving this issue and adds additional benefit of executing all this
mass removals only once instead doing that again and again on each
upgrade.

From now on, please delete deprecated config entries in a regular
upgrade script.

Change-Id: I062b6c9df0f812e3e36cdfed36854dddb045c62d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1269796
Signed-off-by: emesika <emesika@redhat.com>