scheduler: Update database before running integration tests
Apply all database changes in the pre-integration-test phase. This is
by default done with the 'apply' command of the schema.sh script.
Switching branches can lead to incompatible database states. Therefore
it is possible to force maven to recreate the whole database by using
the system property 'engine.db.command' with the value 'recreate'.

Example:

   mvn clean verify -Dengine.db.command=recreate -DskipITs=false

When skipping integrations tests (which is the default), the database
update hook in pre-integration-test  is skipped too.

Change-Id: Ie1fe902255eb4b35dbabb140a5b0c6eb97cba86f
Signed-off-by: Roman Mohr <rmohr@redhat.com>