packaging: setup: support dnf version locking
OTOPI 1.4.0 support DNF package manager.
Upgrade support with DNF package manager requires
changes in versionlock plugin configuration since
DNF and YUM are not sharing the same configuration.

This patch add a new dependency on Fedora adding
the DNF versionlock plugin and configure it as it
does for YUM. Add also the versionlock plugin
configuration to the backupped data.

Change-Id: I16a53c0f39eb8e5a7698ac4fcbba971081da6d24
Bug-Url: https://bugzilla.redhat.com/1270139
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>