engine : Initial size for thin provisioning disk
Add the possibility to specify the volume initial size, for block disk
with thin provisioning.

This option is needed when importing a VM from an external provider
along with its disks.

If one of the VM existing disk is a block disk with thin provisioning,
the new volume needs a starting size as the original disk size.

Change-Id: Id9a970be917de2b3b345b5d95d6e4f04a1423f44
Bug-Url: https://bugzilla.redhat.com/1273284
Signed-off-by: Fred Rolland <frolland@redhat.com>