db: add missing indexes:
This patch includes missing indexes suggested by pgcluu
(http://pgcluu.darold.net/)

PostgreSQL automatically creates indexes on primary keys and unique
constraints, but not on the referencing side of foreign key
relationships.

Since a DELETE of a row from the referenced table or an UPDATE of a
referenced column will require a scan of the referencing table for rows
matching the old value, it is often a good idea to index the referencing
columns.

Change-Id: I09b43335e7cde29235839df00f590fef346d3547
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1270584
Signed-off-by: emesika <emesika@redhat.com>