backend: fix template's disk volume info
Commit 92665e0af829d25273d5d45669847569eae6f5cc introduced the ability
to specify a template disk's format when creating it. However, if the
format was not specified, it defaulted to the active image's format.

This behavior, although not entirely a bad idea (as it prefers thin
provisioning over prealloaction, which may in fact be the best practice
for template disks), is inconsistent with previous versions' behavior,
and could be considered a behavioral regression.

This patch restores the original behavior of matching the template
disk's format with the format of the base image of the disk its derived
from, unless it's explicitly specified, of course.

Change-Id: I7a976c157eb0dbab4fd311a2ae60a57f6c88f4f0
Bug-Url: https://bugzilla.redhat.com/1257240
Signed-off-by: Idan Shaby <ishaby@redhat.com>