engine: Use UUID as content host id on Satellite
There is a difference in the semantic of the ID field under:
 /katello/api/v2/systems/:id resource between Katello to Satellite.

The :id field in Katello represents the UUID of the content host,
however under satellite it is a sequence number, which causes the
requests sent from the engine to fail with unknown resource identifier.
On Satellite (and also on Katello) the value of the UUID field of the
content host can be used to identify it @candlepin.

The solution will be to use the UUID field of the system resource
instead of the ID.

Change-Id: I44ee45ac9770238c8d353ae30bf18a280b92b7ea
Bug-Url: https://bugzilla.redhat.com/1271340
Signed-off-by: Moti Asayag <masayag@redhat.com>