core: Guest's threads-per-core setting added
User can now select threads-per-core for VM in addition to cores-per-socket
to tune simultaneous multithreading.

Guest vCPUs are computed as "sockets X cores X threads" in the GUI.

The "Count Threads As Cores" option relates to host threads only,
the CPUPolicyUnit counts guest's threads per core always as 1 as
a workaround of dynamicaly changed number of threads p. Core on a host.

For x86 hosts it is still reasonable to keep threads-per-core set to 1.
The value set to 8 is recently recommended for high load PPC setup.

Number of sockets set to maximum is still preferred behavior.

Change-Id: I4b49f6f3bcfd2363ec89eee7b1300733ef76691e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1267165
Signed-off-by: Marek Libra <mlibra@redhat.com>