core: Fixed configured range validator
Configured range validator wouldn't add the constraint violation
of range to the context of the error message.
For that, reason when trying to add CPU QoS with out of bound
limit it wouldn't not add the range to the error message.

This problem was fixed by adding the constraint violation to the
context.

Change-Id: I7cefb4d8a114238b86ab187ae4119e6da68b0009
Bug-Url: https://bugzilla.redhat.com/1220110
Signed-off-by: Tomer Saban <tsaban@redhat.com>