engine: Remove snapshot job remains open
The patch fixes few issues with snapshot remove implementation:

1. Commit ad1ac46bf56de9000f6562b7a5d3997326ed9121 has introduced
a change in commands with callback behavior regarding closing
the job. As a result of that patch, the responsibility of closing
the job associated with the command is passed to the callback.

2. The RemoveSnapshotCommand was missing a c'tor which associates
the cached command's context with its new instance while retrieved
by CoCo. Without this c'tor, the command couldn't monitor the job's
progress.

3. Although the command uses compensation, it doesn't introduce the
required c'tor for it.

Change-Id: Ia1331fc864e001a04c0ac1ec6a2b1f500e8fe59e
Bug-Url: https://bugzilla.redhat.com/1257529
Signed-off-by: Moti Asayag <masayag@redhat.com>
(cherry picked from commit 27f8f06c80734547ab0f62ce27c14e64a22de7e7)