core: improve stop of suspended vm
Suspended VM is a VM that is not running but its memory from
previous run was saved so it would be restore next time the
VM will run.

Stop of a suspeneded VM basically means that all we need to
do is to change its status to Down and remove its memory
volumes. However, currently we do not change the status to
Down until the memory volumes are removed.

This patch improves this flow by changing the order of these
operations: first change the status of the VM to Down and
only then remove the memory volumes ('fire and forget').

Change-Id: I76445cf1bd5de5f8129d334d02ed15a415c06848
Signed-off-by: Arik Hadas <ahadas@redhat.com>