core: Removing base templates (backend part)
This patch adds possibility to remote base template even if there are
some sub-templates of that template.

In addition to locking the deleted base template itself, it also shared-
locks the direct successor of deleted base template if there is any.
The lock is acquired in canDoaction() and released at the beginning of
executeCommand() after references to new base template are set.

Change-Id: Iea345cadeb3a02d0f0b82bdd8f4d47a0d6364442
Bug-Url: https://bugzilla.redhat.com/1194989
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>