core: old SetupNetworks fix.
when creating new bond with VLAN using old setupnetworks command, base
nic is calculated and cannot be found in db, because it wasn't created
yet. NPE was produced, when trying to find NetworkAttachment
using ID of inexisting nic. Check was added so code does not try to
find such NetworkAttachment in such scenario. Consequence is, that it
there's QOS defined on NetworkAttachment, it won't be applied. However
this is not regression since this command is deprecated and never
supported overridding QOS.

Bug-Url: https://bugzilla.redhat.com/1271524
Change-Id: I9a2db946352bbd02eaa2dd64b98badd71050b6d3
Signed-off-by: Martin Mucha <mmucha@redhat.com>