engine: Adding NIC to VM fails because boot order is used by another device
When nics are plugged/unplugged, the device boot order on the vm and in the db
can get out of sync. This is not a problem during these operations, as the boot
order is only relevant during vm startup, but it can cause a duplicated boot
order to be submitted when a new device is created. As the boot order is not
relevant during these operations, it is safe to assing no boot order to the
device.

Change-Id: Ib874807e395007a799dc5ff3eb2fd67b218137c6
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1216929