restapi: Avoid NPE in QoS mapper
The mapper is used also when processing an update, and in that case the
"type" can be empty, thus triggering an exception when calling
"toLowerCase". This patch moves the conversion to lower case to the
place where it is actually needed, and where we know that the value
won't be null.

Change-Id: I31ee09a1eb1cd9a706a88fe836b04dfa404841a5
Bug-Url: https://bugzilla.redhat.com/1274187
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>