engine: Remove VM Pool even if VMs are running
Support VM Pool removal even in the case when some VMs from it are
running.

If some of the VMs in the pool are running, RemoveVmPoolCommand stops
them and uses CommandCoordinator to wait until all VMs in the pool will
be down. After that the pool with all VMs is removed as usually.

Change-Id: I778362023e9f03b17e7306425e84245c0b9fcec0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1246886
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>