core, vdsbroker: Refresh volume after extension during Live Merge
When merging a larger (extended) volume into a smaller base volume, the
ExtendImageSize command may be run on a different host than the Merge
command.  In this case the host running Merge may be unaware of the size
change, resulting in a failure.

To prevent this, ExtendImageSize will now send a RefreshVolume request
if a base volume is extended.  The refresh is done on all hosts with the
volume in use because they otherwise have no way of knowing about the
new size.

In addition, we will send a refresh request during live merge recovery
to allow recovery from a failure due to the above size discrepancy when
a merge is retried.

Change-Id: I043c7f14359d1befb3e442958760d2a0dc4c9882
Bug-Url: https://bugzilla.redhat.com/1291569
Signed-off-by: Greg Padgett <gpadgett@redhat.com>