scheduling: Add selector step to the scheduler
The new selector step (and relevant policy units) enhances the
flexibility of our scheduling process.

We used to have just two steps:

- filter out all non-compliant hosts
- compute weight for all remaining hosts

And then a hardcoded routine selected the "best" host
according to the accumulated weights.

The selector replaces this hardcoded logic with a pluggable
policy unit of type SELECTOR. Two example units are provided.

- BasicWeightSelector - copies the old logic
and
- RankSelector - uses rank instead of the raw weights

Change-Id: If63b90c3f5e02d9ccdc77dd920233e3013e933e0
Signed-off-by: Martin Sivák <msivak@redhat.com>