core: Validation of copy Template disk should use image id
The validation for Cinder and LUN disks on copy operation uses the
parameter value of imageGroupId to fetch the disk.

If the operation is being done on a Template, the paramtere of
imageGroupId is not initialized, but instead only with image id is.
That is why the validation does not block the operation of copy Cinder
disk in a Template.

The proposed fix, fetches the disk using the image id when the image
group id is empty or null (which indicates that a copy template disk was
called).

Change-Id: Ibc5f7c228d014a9db2dd74c4edc6d3e13a37a991
Bug-Url: https://bugzilla.redhat.com/1275056
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>