restapi: Populate "placement_policy.hosts" even with only one host
Currently the "placement_policy.hosts" element is populated only if the
VM is pinned to 2 or more hosts. When it is pinned to only one host the
element isn't populated, only "placement_policy.host" is populated. This
patch changes the RESTAPI so that the multiple hosts will be always
populated, even if there is only one host.

Change-Id: I1819a1a2a35ca216071e225c319fb4423d7c3629
Bug-Url: https://bugzilla.redhat.com/1275733
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>