core: Use Java 8 as source and target
This patch changes the POMs so that the new Java 8 capabilities can be
used in all the sub-projects excepts those whose source is used by the
GWT compiler, as it doesn't support Java 8. The patch also fixes some
minor issues in the code that the Java 8 compiler doesn't accept, as it
is more strict in some areas, specially generics.

Change-Id: I0c78347cab79e5d79e096424df94720b4c342edf
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>