restapi: Use JAX-RS Response instead of entity in policy units
Currently the JAX-RS resource interfaces for policy units (filters,
balances and weights) return the entity type as the result of the "add"
methods. All the other resources of the RESTAPI return the JAX-RS
"Reponse" type. This patch changes the policy units resources so that
they return "Response", like the rest of the resources.

Change-Id: I8048c47e75b45f0f5be9e542a4f38e19538772a0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>