restapi: Use "export" instead of "doExport"
The general rule for the names of the JAX-RS interface methods is that
they are named like the action. For example, if the name of the action
is "start" then the name of the method is also "start". The only
exception is when the result is a Java reserved word, like "import", and
in those cases the name is prefixed with "do", resulting in "doImport",
for example. But "export" isn't a reserved word, so in order to improve
consistency this patch renames the "doExport" method to just "export".

Change-Id: I85ecb4c29ff373f9132ef70469170a54d0cadb5c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>