restapi: Use matrix parameters for force remove storage domain
The operation that removes a storage domain supports the "force",
"destroy" and "host" parameters. These parameters were passed to the
"DELETE" method using the representation of the storage domain as the
body:

  DELETE /storagedomains/{storagedomain:id}
  <storage_domain>
    <force>...</force>
    <destroy>...</destroy>
    <host id="...">
      <name>...</name>
    </host>
  </storage_domain>

This is problematic, as the HTTP "DELETE" method shouldn't have a body,
and the representation of the storage domain shouldn't include things
that aren't attributes of the storage domain, rather parameters of the
operation.

This patch replaces the "force", "delete" and "host" attributes by
equivalent matrix parameters, and the operation doesn't now accept a
body. For example, after this patch, the way to delete a storage domain
with the "force" parameter is the following:

  DELETE /storagedomain/{storagedomain:id};host=myhost;force=true

To delete with the "destroy" parameter:

  DELETE /storagedomain/{storagedomain:id};host=myhost;destroy=true

Change-Id: I87d4a9a29fcd9af794ecbedb1b255e676541ee13
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>