packaging: Require slf4j-jdk14 on Fedora
On EL6/7 slf4j package contains all part of slf4j library, but on Fedora
slf4j package was split into several packagesi. Besides api part we
need also jdk14 part (for engine command line tools) which is contained
in slf4j-jdk14 package.

This issue is not visibible on 3.6, because of this dependency:

 ovirt-engine
   -> spring-ldap
        -> springframework-ws
             -> springframework-webmvc
                  -> tiles
                       -> slf4j-jdk14

Engine dependency on spring-ldap was removed on master with removal of
legacy kerbldap provider (engine-manage-domains).

Change-Id: I0448fd16e9ee6b131070ba8d2960c1357ee05744
Bug-Url: https://bugzilla.redhat.com/1264670
Signed-off-by: Martin Perina <mperina@redhat.com>