restapi: Auto-Generate Enums in the API.
This patch adds infrastructure for auto-generating API enums.

'Skeleton' enums, containing only the enum entries, are placed
in 'types' package in 'model' project (alongside the API's
other business entities). A skeleton enum looks like this:

  package types;

  import org.ovirt.api.metamodel.annotations.Type;

  @Type
  public enum DiskFormat {
  	COW, RAW;
  }

Full Enum classes with value() and fromValue() methods are generated
from the skeleton enums and placed in:

 ...definition/target/generated-sources/model/org/ovirt/engine/api/model

directory.

The enums are *not* added to the schema file (api.xsd). The exception
to this rule are three existing enums: ValueType, StatisticType,
StatisticKind, which do exist in the schema for historic reasons.
In the schema file, enums are still represented as strings, for example:

<xs:complexType name="Disk">
  <xs:element maxOccurs="1" minOccurs="0" name="format" type="xs:string"/>
</xs:complexType>

But in the model files, the enums themselves are used (instead of
strings, which is the case today), for example:

@Type
public interface Disk extends Device {
  DiskFormat format();
}

Future patches will delete all existing enums, place skeletons for them
in model project, and replace string references with enum references
where appropriate in model interfaces.

Change-Id: If4643497a5cc0fce8e78d820da8a90c612a3ca21
Signed-off-by: Ori Liel <oliel@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>