core: Quota storage calculating fix for floating disks.
While calculating the quota usage in CalculateStorageUsage psql function,
the calculated "actual size"- which is supposed to include disks
that are not active- DID calculate floating disks in that state.

This patch drops any consideration regarding whether a disk is connected
to a VM or not while calculating the quota usage of inactive disks.

Change-Id: Ib417215b2db6e0ee5ae3acad17b7dcf77b175cf6
Bug-Url: https://bugzilla.redhat.com/1225484
Signed-off-by: Amit Aviram <aaviram@redhat.com>