core: when starting Live Merge, check for NOT_STARTED child commands
When a Live Merge is started, the RemoveSnapshotCommandCallback does not
look for child commands with the NOT_STARTED command status.  This
causes a race where the child commands may not be detected properly by
the parent, causing the actual merge process to be skipped.

This adds the needed check to the command callback, eliminating the race
condition and ensuring the merge will occur.

Change-Id: If47d26d8bedaf0bb4a39cae472533399178b4fdd
Bug-Url: https://bugzilla.redhat.com/1280160
Signed-off-by: Greg Padgett <gpadgett@redhat.com>