restapi: Generate Action XML types
Currently the Action XML type is defined in the XML schema, in the
api.xsd file. This type is basically the union of two groups of
elements: ActionParameterGroup and ActionResponseGroup. The first is the
union of all the input parameters of all the action methods. The second
is the union of all the output parameters of all the action methods.

This patch removes all the Action related XML types from the api.xsd
file and changes the schema generator so that it generates them from the
model. The specifications of the parameters are also moved to the
corresponding service interfaces.

For example, the action to migrate a virtual machine uses a "force"
parameter that is currently defined in the XML schema as follows:

  <!-- should migration be forced if VM is pinned to host? -->
  <xs:element name="force" type="xs:boolean" minOccurs="0"/>

This patch moves it to the definition of the virtual machine service,
the VmService interface, where it is specified as follows:

  @Type
  public interface VmService {

    interface Migrate {
      /**
       * Indicates if migration should be forced if VM is
       * pinned to host.
       */
      @In Boolean force();
    }

  }

The schema generator collects the parameters of all the services and
generates the ActionParameterGroup and ActionResponseGroup from that.

Change-Id: I581f98794c21e6c03df897d3415dfbc55d1820a8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>