core: show snapshot name properly after live merge of snapshot disk
After Live Merge of a snapshot disk (Storage > Disk Snapshots > Remove),
the audit log shows the snapshot name of '<UNKNOWN>'.  The problem is
that the image record is used to look up the name, and this is not
available after the merge operation is done.

Instead, cache the snapshot name in the parameters and refer to this
after the merge operation is complete.

Change-Id: I253806b89f31cb31e0a145dfe22a6f3ba30579f8
Bug-Url: https://bugzilla.redhat.com/1204609
Signed-off-by: Greg Padgett <gpadgett@redhat.com>