core: separate tests differing with bond mode
I have some issues with this test in following patch(do not pass whole
Bond from UI, but only required data), and it would be harder to
fix that problem differently than it's done in this patch.

There are several tests, which tests same stuff for all bond options.
These tests saves some computational time by reinitializing only bond
instance between tests, but this is cause of my problems, since I'm
replacing datatype of parameters.newOrModifiedBonds, so I'd have to
update inside of loop not only bond, but also all bonds in given
collection. So I opted to completely separate test for each bond mode.

This way is of course less
performant(not significantly though), but does not assume correct
behavior of validator being tested. Before this modification, tested
subject had to 'cooperate' and neither do touch 'bond' passed in
parameters nor to store data from it in any its [validators] inner
structures. Doing so whould probably cripple test, and test
should probably not assume this is not happening if it's possible to
write it like that.

I also introduced ErrorCollector, which will report all failing
bondmodes and not just first one.

Change-Id: I654b60f8456a16f5925658a84f8b44d5c7a7b5e4
Signed-off-by: Martin Mucha <mmucha@redhat.com>