restapi: Handle incorrect user and group identifiers
Currently the RESTAPI doesn't check the format of user or group
identifiers provided by the caller, it just tries to convert them
assuming that they are hexadecimal characters. If this fails the
exception isn't handled properly and stack trace is returned to the
user. This patch changes the relevant resources so that they handle the
exception and return the expected HTTP response codes.

Change-Id: I1bf0139fd329f63e542d24b8d9f0115e8e7b6c80
Bug-Url: https://bugzilla.redhat.com/1281803
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 549ce58658212b749e02fdfe59beace518ab598a)