core: StorageHandlingCommandBase#isDomainExistsInDiskDescription
Simplified method call by using Objects.toString with a default instead
of explicitly checking for nulls.
Besides shortening the code and making it more fluid to read, this patch
offers a minor performance improvement as
map.get(OvfInfoFileConstants.Domains) is now only called once, but it's
doubtful that this improvement will be noticeable in any reasonable
(or, frankly, unreasonable) setup.

Change-Id: I3de0cf7eae0926d650f2c6b8064c30c03a919cca
Signed-off-by: Allon Mureinik <amureini@redhat.com>