frontend: new VM from template blank value in target sd list
When creating a VM from a template, one can choose on which storage
domain to store each new VM's disk.
Given the list of storage domains where we can do that, we used to
select the first storage domain by default from disk.getStorageIds().
The problem was that if this storage domain was not in the list of
storage domains where we can store the new VM's disk, null was selected,
creating a blank option in the list and blocking the whole command from
executing.

This patch uses Linq.IdsPredicate instead of Linq.StoragePredicate.
Thus, now we go over the list of storage domains, check whether each
storage domain exists in disk.getStorageIds(), and select it if it does.
If there's no match, the first storage domain in the list will be
selected by default (by using Linq.firstOrDefault instead of
Linq.firstOrNull), and the command will fail the canDoAction with an
appropriate error stating that there's no active storage domain the disk
can be read from.

Change-Id: Ic9628c9404030f38cae8e7bb13edca4a299594db
Bug-Url: https://bugzilla.redhat.com/1282085
Signed-off-by: Idan Shaby <ishaby@redhat.com>