scheduler: smarter thread management in thread pool
SimpleThreadPool create all the threads in advance and keep them
running/waiting all the time but we could be more flexible with thread
management and deallocate them once not needed.

In order to manage thread well we reimplemented quartz ThreadPool
interface and used ThreadPoolExecutor as implementation which improves
thread management significantly.


Change-Id: I6776bbfcbf3c069295fda300214bfdd02375bb3c
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>